// libs/domains/iam/iam-domain/src/lib/value-objects/role-scope.enum.ts
/**
 * Đại diện cho phạm vi của vai trò trong hệ thống.
 * 
 * @since 1.0.0
 */
export enum RoleScope {
  /**
   * Vai trò nội bộ Ecoma
   */
  INTERNAL = 'Internal',

  /**
   * Vai trò trong phạm vi một tổ chức khách hàng
   */
  ORGANIZATION = 'Organization'
} 